# users/router.ts
import { Router, Request, Response, Handler } from "express";
import { inviteUser, user_list, edit_user as edit_user, user_status, user_login, userInviteResend, RegisterUser, userDetails, userRoles, userCapabilities, forgotPassword, setNewPassword, createGroup, editGroup, groupList, groupStatus, groupDetail, addMember, removeMembers, userSuggestions, otpVerification, userInformation, changeEmailInfo, getUserDetail } from "./module";
import { authenticate } from "../utils/utils";
import { NextFunction } from "connect";
var multer = require('multer');
import { readFileSync } from "fs";
import { join } from "path";
import { changePasswordInfo, uploadPhoto } from "../utils/users";
import { uploadToFileService } from "../documents/module";
import { get as httpGet } from "http";
import { get as httpsGet } from "https";
import { FILES_SERVER_BASE } from "../utils/urls";
import { APIError } from "../utils/custom-error";
const router = Router();

//  Invite User
router.post('/create', authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await inviteUser(req.body, res.locals.user));
    } catch (err) {
        next(new APIError(err.message));
    };
});

// Register User
router.post("/register/:token", async (req: Request, res: Response, next: NextFunction) => {
    try {
        if(req.body.upfile){
            const imageObj: any = await uploadToFileService(req)
            req.body.profilePic = JSON.parse(imageObj).id
        }
        res.status(200).send(await RegisterUser(req.body, req.params.token))
    } catch (err) {
        next(new APIError(err.message));
    };
});


//  Resend invite link
router.get('/invite/resend/:role/:id', authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        const { role, id } = req.params
        res.status(200).send(await userInviteResend(id, role));
    } catch (err) {
        next(new APIError(err.message));
    };
});

//  user list
router.get('/list', authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        req.query.page = req.query.page || 1;
        req.query.limit = 50;
        res.status(200).send(await user_list(req.query, res.locals.user._id, req.query.page, req.query.limit));
    } catch (err) {
        next(new APIError(err.message));
    };
});

router.get(`/detail/:id`, authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await getUserDetail(req.params.id));
    } catch (err) {
        next(new APIError(err.message));
    };
});
//  Edit User
router.post('/edit/:id', authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        if(req.body.upfile){
            const imageObj: any = await uploadToFileService(req)
            req.body.profilePic = JSON.parse(imageObj).id
        }
        res.status(200).send(await edit_user(req.params.id, req.body, res.locals.user));
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  change status user
router.put('/status/:id', authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await user_status(req.params.id, res.locals.user));
    } catch (err) {
        next(new APIError(err.message));
    };
});

//  login user
router.post('/email/login', async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await user_login(req.body));
    } catch (err) {
        next(new APIError(err.message));
    };
});

router.get(`/getImage/:userId`, async (request, response, next) => {
    try {
        const userDetail = await getUserDetail(request.params.userId)
        const req = (FILES_SERVER_BASE as string).startsWith("https") ? 
        httpsGet(`${FILES_SERVER_BASE}/files/${userDetail.profilePic}`, (res: any) => {
            response.setHeader('Content-disposition', 'inline');            
            response.setHeader('Content-type',res.headers['content-type'])
            res.pipe(response);
        }) : httpGet(`${FILES_SERVER_BASE}/files/${userDetail.profilePic}`, (res: any) => {
            response.setHeader('Content-disposition', 'inline');            
            response.setHeader('Content-type',res.headers['content-type'])
            res.pipe(response);
        });
        req.on('error', (e: Error) => {
            next(e);
        });
        req.end();
    } catch (err) {
        next(new APIError(err.message));
    };
});
// Get user Details
router.get("/me", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await userDetails(res.locals.user._id));
    } catch (err) {
        next(new APIError(err.message));
    };
});

// Get user roles
router.get("/me/role", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await userRoles(res.locals.user._id));
    } catch (err) {
        next(new APIError(err.message));
    };
});

// Get user capabilities
router.get("/me/capabilities", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await userCapabilities(res.locals.user._id));
    } catch (err) {
        next(new APIError(err.message));
    };
});

// Forgot Password
router.post("/forgot", async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await forgotPassword(req.body));
    } catch (err) {
        next(new APIError(err.message));
    };
});

router.post("/forgot/verify", async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await otpVerification(req.body));
    }
    catch (err) {
        next(new APIError(err.message));
    }
});

// Set New Password
router.post("/forgot/setPassword", async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await setNewPassword(req.body));
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  Add Group
router.post("/group/create", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await createGroup(req.body))
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  List Group
router.get("/group/list", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await groupList())
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  Edit Group
router.put("/group/:id/edit", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await editGroup(req.body, req.params.id))
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  edit status of group
router.put("/group/:id/status", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await groupStatus(req.params.id))
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  Add Member
router.post("/group/:id/member/add", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await addMember(req.params.id, req.body.users));
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  Remove Member
router.post("/group/:id/member/remove", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await removeMembers(req.params.id, req.body.users));
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  Get User suggestion
router.get("/suggestion", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await userSuggestions(req.query.search));
    } catch (err) {
        next(new APIError(err.message));;
    };
});

//  Group Details
router.get("/group/:id", authenticate, async (req: Request, res: Response, next: NextFunction) => {
    try {
        res.status(200).send(await groupDetail(req.params.id));
    } catch (err) {
        next(new APIError(err.message));;
    };
});
router.get("/countryCodes", async (req, res, next) => {
    try {
        res.status(200).send(JSON.parse(readFileSync(join(__dirname, "..", "utils", "country_codes.json"), "utf8")))
    } catch (err) {
        next(new APIError(err.message));
    }
})

router.get("/userInfo/:id", authenticate, async (req, res, next) => {
    try {
        res.status(200).send(await userInformation(req.params.id));
    } catch (err) {
        next(new APIError(err.message));
    }
})
router.post("/changePassword",authenticate, async (req, res, next) => {
    try {
        res.status(200).send(await changePasswordInfo(req.body,res.locals.user._id));
    } catch (err) {
        next(new APIError(err.message));
    }
})
router.post('/changeEmail',authenticate,async (req,res,next)=>{
    try{
        res.status(200).send(await changeEmailInfo(req.body,res.locals.user._id));
    }
    catch(err){
        next(new APIError(err.message));;
    }
})
export = router;